Se utiliza una sola función para todos los endpoint de connect
