+++ back/routes/auth/facebook.js
@@ -1,9 +1,11 @@
 const express = require('express');
 const passport = require('passport');
 const router = express.Router();
+const { passport_connect } = require('./utils');
 const strategy_name = 'facebook';
+const strategy_scope = ['email'];
 
-router.get('/facebook/auth', passport.authenticate(strategy_name, { session:false, scope: ['email']}));
+router.get('/facebook/auth', passport.authenticate(strategy_name, { session:false, scope: strategy_scope}));
 
 router.get('/facebook/connect', function (req, res, next) {
   /* Connects the current user account with Google. */
@@ -12,12 +14,10 @@ router.get('/facebook/connect', function (req, res, next) {
 
   console.log("New request GET to /facebook/connect");
 
-  const user_id = 1;  // TODO: get the user id from the token
-  const state = `${user_id}`;  // state must be string
+  // We supose that the middleware defines the req.user object
+  req.user = {id: 1,}
 
-  // redirect to facebook to authenticate
-  passport.authenticate(strategy_name, { session:false, scope: ['email'], state: state })(req, res, next);
-
+  passport_connect(strategy_name, strategy_scope, req, res, next);
 });
 
 router.get('/facebook/callback', passport.authenticate(strategy_name, {  session:false, failureRedirect: '/failed' }),
