+++ back/routes/auth/google.js
@@ -1,23 +1,11 @@
 const express = require('express');
 const passport = require('passport');
 const router = express.Router();
+const { passport_connect } = require('./utils');
 const strategy_name = 'google';
 const strategy_scope = ['profile', 'email'];
 
 router.get('/google/auth', passport.authenticate(strategy_name, { session:false, scope: strategy_scope}));
-
-function passport_connect(strategy_name, strategy_scope, req, res, next) {
-
-  const user_id = req.user.id;
-  const state = `${user_id}`;  // state must be string
-
-  // redirect to strategy to authenticate
-  let passport_authenticate = passport.authenticate(
-    strategy_name, { session:false, scope: strategy_scope, state: state }
-  );
-  passport_authenticate(req, res, next);
-}
-
 
 router.get('/google/connect', function (req, res, next) {
   /* Connects the current user account with Google. */
